Remove undershoot and fix scss build.
For some reason, the scss file for "_drawing.scss" wasn't
building, but I've fixed it now. I might be able to get
rid of my variable I added, too.

For some reason, the bottom of some UI elements doesn't
have an edge anymore. Could be the 1px thing with line 372.