Improve titlebar contrast for light theme.
The light theme now uses a value of 80,80,80 for the
focused titlebar text; and unfocused window titlebar
backgrounds are now 80% for their alpha value.

The exact color for focused titlebar text may change in
the future, but this should be ok for now.

Unfocused window titlebars still use the default Mint-Y
theme values for the color of the text and background,
with the exception of their alpha value.